Refactor to use context module
