+++ packages/hooks/src/decorator.ts
@@ -1,6 +1,6 @@
 import { functionHooks } from './function';
 import { setManager, HookOptions, convertOptions } from './base';
-import { contextProperties } from './hooks';
+import { properties } from './context';
 
 export const hookDecorator = (managerOrMiddleware?: HookOptions) => {
   const wrapper: any = (_target: any, method: string, descriptor: TypedPropertyDescriptor<any>): TypedPropertyDescriptor<any> => {
@@ -18,7 +18,7 @@ export const hookDecorator = (managerOrMiddleware?: HookOptions) => {
       throw new Error(`Can not apply hooks. '${method}' is not a function`);
     }
 
-    manager._middleware.unshift(contextProperties({ method }));
+    manager._middleware.unshift(properties({ method }));
     descriptor.value = functionHooks(fn, manager);
 
     return descriptor;
