+++ packages/hooks/test/function.test.ts
@@ -7,9 +7,7 @@ import {
   NextFunction,
   setMiddleware,
   functionHooks,
-  contextParams,
-  contextProperties,
-  contextDefaults
+  context
 } from '../src';
 
 describe('functionHooks', () => {
@@ -18,7 +16,7 @@ describe('functionHooks', () => {
   };
 
   it('returns a new function, registers hooks', () => {
-    const fn = hooks(hello, []);
+    const fn = hooks(hello, middleware());
 
     assert.notDeepEqual(fn, hello);
     assert.ok(getManager(fn) !== null);
@@ -161,10 +159,10 @@ describe('functionHooks', () => {
 
   it('chains context initializers', async () => {
     const first = hooks(hello, middleware([
-      contextProperties({ testing: ' test value' })
+      context.properties({ testing: ' test value' })
     ]));
     const second = hooks(first, middleware([
-      contextParams('name'),
+      context.params('name'),
       async (ctx, next) => {
         ctx.name += ctx.testing;
         await next();
@@ -178,7 +176,7 @@ describe('functionHooks', () => {
 
   it('creates context with params and converts to arguments', async () => {
     const fn = hooks(hello, middleware([
-      contextParams('name'),
+      context.params('name'),
       async (ctx, next) => {
         assert.equal(ctx.name, 'Dave');
 
@@ -193,8 +191,8 @@ describe('functionHooks', () => {
 
   it('assigns props to context', async () => {
     const fn = hooks(hello, middleware([
-      contextParams('name'),
-      contextProperties({ dev: true }),
+      context.params('name'),
+      context.properties({ dev: true }),
       async (ctx, next) => {
         assert.equal(ctx.name, 'Dave');
         assert.equal(ctx.dev, true);
@@ -219,7 +217,7 @@ describe('functionHooks', () => {
     };
 
     const fn = hooks(hello, middleware([
-      contextParams('name'),
+      context.params('name'),
       modifyArgs
     ]));
 
@@ -237,7 +235,7 @@ describe('functionHooks', () => {
   it('can take and return an existing HookContext', async () => {
     const message = 'Custom message';
     const fn = hooks(hello, middleware([
-      contextParams('name'),
+      context.params('name'),
       async (ctx, next) => {
         assert.equal(ctx.name, 'Dave');
         assert.equal(ctx.message, message);
@@ -320,8 +318,8 @@ describe('functionHooks', () => {
 
   it('creates context with default params', async () => {
     const fn = hooks(hello, middleware([
-      contextParams('name', 'params'),
-      contextDefaults(() => {
+      context.params('name', 'params'),
+      context.defaults(() => {
         return {
           name: 'Bertho',
           params: {}
